avoid rework, prune graph and stop watching on relevant changes
