+++ packages/core/core-v2/src/Parcel.js
@@ -3,7 +3,7 @@
 const path = require('path');
 const { AbortController } = require('abortcontroller-polyfill/dist/cjs-ponyfill');
 const PQueue = require('p-queue');
-const Graph = require('./Graph');
+const AssetGraph = require('./AssetGraph');
 const TransformerRunner = require('./TransformerRunner');
 const ResolverRunner = require('./ResolverRunner');
 const BundlerRunner = require('./BundlerRunner');
@@ -22,7 +22,7 @@ class Parcel {
   }) {
     this.rootDir = process.cwd();
 
-    this.graph = new Graph({ entries, rootDir: this.rootDir });
+    this.graph = new AssetGraph({ entries, rootDir: this.rootDir });
     this.watcher = cliOpts.watch ? new Watcher() : null;
     this.queue = new PQueue({
       autoStart: false,
@@ -57,6 +57,7 @@ class Parcel {
 
   async build({ signal }) {
     let graph = await this.completeGraph({ signal });
+    // await graph.dumpGraphViz();
     let { bundles } = await this.bundle(graph);
     await this.package(bundles);
   }
@@ -85,8 +86,9 @@ class Parcel {
   async resolve(depNode, { signal }) {
     // console.log('resolving depNode', depNode)
     let resolvedPath = await this.resolverRunner.resolve(depNode.value);
-    if (!signal.aborted) {
-      let file = { filePath: resolvedPath };
+
+    let file = { filePath: resolvedPath };
+    if (!signal.aborted && !this.graph.hasFileNode(file)) {
       let fileNode = this.graph.addFileNode(depNode, file);
       this.queue.add(() => this.transform(fileNode, { signal }));
       if (this.watcher) this.watcher.watch(resolvedPath);
@@ -96,15 +98,35 @@ class Parcel {
   async transform(fileNode, { signal }) {
     // console.log('transforming fileNode', fileNode)
     let { children: childAssets } = await this.transformerRunner.transform(fileNode.value);
-
     if (!signal.aborted) {
+      let assetNodesToRemove = fileNode.fromEdges.map(edge => this.graph.nodes.get(edge.to));
+      let depNodesToRemove = [];
+
       for (let asset of childAssets) {
         let assetNode = this.graph.addAssetNode(fileNode, asset);
+        assetNodesToRemove = assetNodesToRemove.filter(node => node === assetNode);
+
+        let assetDepNodes = assetNode.fromEdges.map(edge => this.graph.nodes.get(edge.to))
+        depNodesToRemove = depNodesToRemove.concat(assetDepNodes);
 
         for (let dep of asset.dependencies) {
           dep.sourcePath = fileNode.value.filePath; // ? Should this be done elsewhere?
-          let depNode = this.graph.addDependencyNode(assetNode, dep);
-          this.queue.add(() => this.resolve(depNode, { signal }));
+          if (!this.graph.hasDependencyNode(dep)) {
+            let depNode = this.graph.addDependencyNode(assetNode, dep);
+            this.queue.add(() => this.resolve(depNode, { signal }));
+          } else {
+            let depNode = this.graph.getDependencyNode(dep);
+            depNodesToRemove = depNodesToRemove.filter(node => node === depNode);
+          }
+        }
+      }
+
+      let nodesToRemove = [...assetNodesToRemove, ...depNodesToRemove];
+      for (let node of nodesToRemove) {
+        let { nodes: prunedNodes } = this.graph.prune(node);
+        let prunedFiles = prunedNodes.filter(node => node.type === 'file').map(node => node.value);
+        for (let file of prunedFiles) {
+          if (this.watcher) this.watcher.unWatch(file.filePath);
         }
       }
     }
